requirements.txt
pycaret==3.2.0
streamlit
Pillow